+++ back/src/main/java/com/openclassrooms/starterjwt/services/UserService.java
@@ -1,8 +1,14 @@
 package com.openclassrooms.starterjwt.services;
 
+import com.openclassrooms.starterjwt.exception.NotFoundException;
+import com.openclassrooms.starterjwt.exception.UnauthorizedException;
 import com.openclassrooms.starterjwt.models.User;
 import com.openclassrooms.starterjwt.repository.UserRepository;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.stereotype.Service;
+
+import java.util.Objects;
 
 @Service
 public class UserService {
@@ -13,6 +19,17 @@ public class UserService {
     }
 
     public void delete(Long id) {
+        User user = findById(id);
+        if (user == null) {
+            throw new NotFoundException();
+        }
+
+        UserDetails userDetails = (UserDetails) SecurityContextHolder.getContext().getAuthentication().getPrincipal();
+
+        if (!Objects.equals(userDetails.getUsername(), user.getEmail())) {
+            throw new UnauthorizedException();
+        }
+
         this.userRepository.deleteById(id);
     }
